+++ src/types/dx-data-grid.ts
@@ -37,6 +37,7 @@ export type DxDataGridColDefClickArgDetails = {
 export type DxDataGridColDef = {
   field: string;
   headerName?: string;
+  subtitle?: (data: unknown, column: DxDataGridColDef) => string | undefined,
   hideSortIcon?: boolean;
   editIcon?: boolean;
   editLink?: string