+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added ARIA accessibility tests for `dx-preview` and `dx-dialog` components
 - Added `focusOnLoadingContainer()` public method in `dx-data-grid-generic` component
 - Added `focusDialog()` public method in `dx-dialog` component
+- Added `subtitle` part to `dx-data-grid-generic` component.
 
 ### Fixed
 - Fixed screen reader accessibility in `dx-preview` component by adding proper ARIA labels, roles, and attributes
@@ -18,6 +19,7 @@
 - Fixed keyboard accessibility for `dx-dialog` close button
 - Improved accessibility of `dx-breadcrumbs` component for screen readers
 - Improved accessibility of `dx-dialog` component for screen readers
+- Fix `dx-preview` component bug of previewing the same item will open the first index item in preview.
 
 ### Changed
 - Changed `dx-preview` backdrop to use `role="presentation"` for better screen reader experience